requirements.txt
flask==3.1.1
flask-bcrypt==1.0.1
flask-cors==6.0.1
flask-mail==0.10.0
flask-marshmallow==1.3.0
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.10
pyjwt==2.10.1
sqlalchemy==2.0.41
flask-talisman==1.1.0
flask-limiter==3.12